% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancemb-package.R
\docType{package}
\name{ancemb-package}
\alias{ancemb-package}
\alias{ancemb}
\title{ancemb: Bayesian Ancestral Reconstruction of Protein Embeddings}
\description{
Provides functions to reconstruct ancestral protein embeddings from
extant (tip) embeddings and a phylogenetic tree. Models are fit with
pre-compiled Stan models using Brownian Motion or Ornstein-Uhlenbeck
evolutionary processes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alrobles.github.io/ancemb/}
  \item Report bugs at \url{https://github.com/alrobles/ancemb/issues}
}

}
\author{
\strong{Maintainer}: Angel Luis Robles Fernandez \email{a.l.robles.fernandez@gmail.com} (\href{https://orcid.org/0000-0002-4674-4270}{ORCID})

Authors:
\itemize{
  \item Angel Luis Robles Fernandez \email{a.l.robles.fernandez@gmail.com} (\href{https://orcid.org/0000-0002-4674-4270}{ORCID})
}

}
\keyword{internal}
