% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{bm_covariance_r}
\alias{bm_covariance_r}
\title{Brownian Motion covariance for all node pairs}
\usage{
bm_covariance_r(tree, rate)
}
\arguments{
\item{tree}{A `phylo` object.}

\item{rate}{Numeric vector of length 1 or `D`.}
}
\value{
Matrix or 3D array. If `rate` is length 1, a matrix is returned;
  otherwise an `n_total x n_total x D` array.
}
\description{
Brownian Motion covariance for all node pairs
}
\keyword{internal}
