% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{build_all_node_covariance_r}
\alias{build_all_node_covariance_r}
\title{Build the phylogenetic shared-height matrix for all nodes}
\usage{
build_all_node_covariance_r(tree)
}
\arguments{
\item{tree}{A `phylo` object.}
}
\value{
A symmetric matrix with `n_tips + n_nodes` rows/columns.
}
\description{
Returns the matrix `M` where `M[i,j]` is the height of the MRCA of nodes
`i` and `j`. This is the generalisation of `ape::vcv` to all nodes.
}
\keyword{internal}
