% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{cosine_sim_r}
\alias{cosine_sim_r}
\title{Cosine similarity between two vectors}
\usage{
cosine_sim_r(a, b)
}
\arguments{
\item{a, b}{Numeric vectors.}
}
\description{
Cosine similarity between two vectors
}
\keyword{internal}
