% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ancestral.R
\name{.detect_family}
\alias{.detect_family}
\title{Detect model family from a Stan fit}
\usage{
.detect_family(fit)
}
\arguments{
\item{fit}{A `CmdStanMCMC` object.}
}
\value{
Character `"bm"` or `"ou"`.
}
\description{
Uses the presence of `rate[1]` (BM) or `sigma2[1]` (OU) in the posterior
draws.
}
\keyword{internal}
